+++ examples/widgets/formButton/Schema.js
@@ -1,0 +1,3 @@
+import { schema } from 'epage-core'
+
+export default schema.base.formButton