+++ examples/widgets/formButton/Setting.js
@@ -1,0 +1,3 @@
+import { widget } from 'epage-core'
+
+export default []