+++ examples/widgets/richText/index.js
@@ -1,0 +1,3 @@
+export { default as View } from './View'
+export { default as Setting } from './Setting'
+export { default as Schema } from './Schema'