+++ src/widget/base/button/index.js
@@ -1,4 +1,0 @@
-import { schema } from 'epage-core'
-
-export { default as Setting } from './Setting'
-export const Schema = schema.base.button