+++ src/widget/components/index.js
@@ -1,3 +1,0 @@
-export { default as DataSource } from './data-source'
-export { default as FieldTip } from './field-tip'
-export { default as MoveBtn } from './move-btn'