+++ src/widget/setting.js
@@ -1,9 +1,0 @@
-import SettingForm from './setting-form'
-import settingExtend from './extends/setting'
-import * as components from './components'
-
-export {
-  SettingForm,
-  settingExtend,
-  components
-}